refactor(#54): cleaning up helpers
